requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
certifi==2023.7.22
charset-normalizer==3.3.2
dnspython==2.4.2
ffmpeg-python==0.2.0
hachoir==3.2.0
idna==3.4
motor==3.3.2
Pillow==10.0.1
psutil==5.9.6
pymongo==4.5.0
Pyrogram==2.0.106
pyromod==1.5
python-dotenv==1.0.0
requests==2.31.0
tenacity==8.2.3
TgCrypto==1.2.5
urllib3==2.0.7
humanize==4.8.0
httpx>=0.24.0
aiodns>=3.0.0
colorama>=0.4.6
rich>=13.0.0
asyncio-throttle==1.0.2
colorlog==6.8.0
python-magic>=0.4.27
aiodns
aiofiles
aiohttp
asyncio-throttle
colorama
colorlog
ffmpeg-python
hachoir
httpx
humanize
motor
Pillow
psutil
pymongo
pyrogram
pyromod
python-dotenv
requests
rich
tenacity
TgCrypto
